+++ core/src/main/java/org/nervos/ckb/unlocker/ScriptType.java
@@ -1,7 +1,7 @@
-package org.nervos.ckb.signature;
+package org.nervos.ckb.unlocker;
 
 // Todo: there is another org.nervos.indexer.model.ScriptType. To avoid circular dependency I create
-// a new one. Need refactor
+// a new one. Need refactor.
 public enum ScriptType {
   LOCK,
   TYPE
