+++ core/src/main/java/org/nervos/ckb/unlocker/TransactionWithScriptGroups.java
@@ -1,4 +1,4 @@
-package org.nervos.ckb.signature;
+package org.nervos.ckb.unlocker;
 
 import java.util.ArrayList;
 import java.util.List;
